Better optimisation opportunities in generated query.
The SPARQL query generated from a DSQPI query does not work as smoothly with ARQ
optimisations as we would like, in particular it does not put grounded ?item queries
preferentially at the top and it puts filters all the way at the bottom, in at least
some circumstances. Experimentally we are splitting the where-clause elements into
more slots (search, grounded, ungrounded, filters, optionals, bindings) and allocating
elements in the appropriate slot before rendering those slots in a standard order.