Preliminary tests for clause ordering.
Preliminary tests for ordering have terms which encode a piece of JSON DSAPI query
and a fragment of expected SPARQL, and checks that given two terms that the
fragments appear in the correct order, viz, the first term should come ealier
than the second.

This form doesn't really catch all the ordering information so the next version
will explicitly specify the SPARQL fragments and their ordering.