Improved testing for clause ordering.
The tests for clause ordering now take the form 'for THIS query, the following fragments
should appear in this order in the generated query.'. For the moment we ignore the
issue that within a section of the generated query theser isn't a well-defined
order -- we'll just work around thta for now.