Improve handling of filters by incorporating aspect objects.
Filter elements used to contain an SQ_Variable which was the variable to be bound to the
aspect's value. However it turns out to be helpful if the filter knows whether it
belongs to an optional aspect or not, since if it belongs to an optional it is
better to leave it till the end of the query. So now the filter includes the aspect
it is filtering and generates its SQ_Variable on demand.