Suppressed diagnostic outputs and one test.
The one test that checks the number of elements that is returned
from a quary fails because the query can produce a different number of results
depending on ... something we're not sure of at the moment. This is probably
a multiple-values issue and will be revisited shortley now the big engineering
push is nearly over.