SearchSpec now early binds ?item to text query result.
The SearchSpec generated an early ?X text:query (Objects) triple
but did not bind ?item to the value of the searched property
immediately. As a result, literal-bound triples appeared between
the search and the search binding. This generated a drastic
performance hit.

Fixed by adding a methods to put triples explicitly in the
early query section and invoking that from SearchSpec.